snapd: disable Go shared runtime for x86
snapd crashes with a SIGSEGV on x86 if shared runtime (libstd.so) is enabled.
The problem was reproduced on Ubuntu 16.04 and only on i386. Going though
numerous Go versions, 1.9 was the first one where the problem was introduced.
Bisecting further, the problem was narrowed down to commit [1]

Until we have a better solution, temporarily disable shared runtime for x86.

See the forum thread [2] for more details.

[1]. https://github.com/golang/go/commit/4808fc444307fa683bf3df6d55f9ad1828891a36
[2]. https://forum.snapcraft.io/t/yocto-rocko-core-snap-panic/3261

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>